refactor(editor): Modularize PropertiesPanel and reorganize component structure
BREAKING CHANGES:
- Moved editor components into feature-based folders
- Updated import paths for relocated components

REFACTORING:
PropertiesPanel Modularization:
- Split monolithic PropertiesPanel.tsx (631 lines) into focused components
- Created EmptyPropertiesState.tsx for empty state display
- Created TextStyleSection.tsx (438 lines) for text styling controls
- Created TransformSection.tsx (89 lines) for transform/timing controls
- Reduced main PropertiesPanel.tsx to 139 lines (78% reduction)
- Improved maintainability and testability

Component Organization:
- Reorganized editor components into feature-based folders:
  * preview/ - Preview and playback components
  * properties/ - Properties panel sections
  * timeline/ - Timeline components (added index.ts)
  * media-panel/ - Media library components
  * media-tabs/ - Media tab system
  * viewport/ - Viewport controls and overlays
  * transform/ - Transform system components

- Moved components to appropriate folders:
  * PreviewPanel, PreviewTransport, GPUPreview → preview/
  * SourcePreview, TextSourcePreview → preview/
  * EnhancedMediaPanel, AudioWaveform → media-panel/
  * ViewportControls → viewport/
  * Test files co-located with components

Index Files & Clean Imports:
- Added index.ts to all feature folders for clean exports
- Created main editor/index.ts for centralized exports
- Simplified import statements across the codebase

Documentation:
- Added comprehensive README.md explaining structure
- Added STRUCTURE.md with visual diagrams and data flow
- Documented component responsibilities and import patterns
- Included refactoring metrics and maintenance guidelines

BENEFITS:
✅ 78% reduction in PropertiesPanel.tsx size
✅ Clear separation of concerns
✅ Improved code maintainability
✅ Better component discoverability
✅ Easier testing and debugging
✅ Scalable folder structure
✅ Co-located tests with components
✅ Clean, consistent import patterns

FILES CHANGED:
- Modified: PropertiesPanel.tsx, EditorLayout.tsx, Timeline.tsx
- Created: 3 new property components + 8 index files + 2 docs
- Moved: 11 components to feature folders
- Organized: Test files co-located with components
