+++ src/components/editor/index.ts
@@ -1,0 +1,32 @@
+/**
+ * Video Editor Components
+ *
+ * Main entry point for all video editor components.
+ * Components are organized by feature area for better maintainability.
+ */
+
+// Main Layout
+export { EditorLayout } from "./EditorLayout";
+export { TopBar } from "./TopBar";
+
+// Preview Components
+export * from "./preview";
+
+// Properties Panel
+export { PropertiesPanel } from "./PropertiesPanel";
+export * from "./properties";
+
+// Timeline Components
+export * from "./timeline";
+
+// Media Panel
+export * from "./media-panel";
+
+// Media Tabs
+export * from "./media-tabs";
+
+// Viewport Components
+export * from "./viewport";
+
+// Transform Components
+export * from "./transform";
