+++ src/components/editor/media-panel/AudioWaveform.tsx
