+++ src/components/editor/media-panel/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Media Panel Components
+ *
+ * Components for the media panel that displays available
+ * media assets and audio waveforms.
+ */
+
+export { EnhancedMediaPanel } from "./EnhancedMediaPanel";
+export { AudioWaveform } from "./AudioWaveform";
