+++ src/components/editor/properties/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Properties Panel Components
+ *
+ * Modular components for the properties panel that displays
+ * clip properties and styling options.
+ */
+
+export { EmptyPropertiesState } from "./EmptyPropertiesState";
+export { TextStyleSection } from "./TextStyleSection";
+export { TransformSection } from "./TransformSection";
