+++ src/components/editor/timeline/index.ts
@@ -1,0 +1,20 @@
+/**
+ * Timeline Components
+ *
+ * Components for the video timeline, including tracks, clips,
+ * playhead, rulers, and drag-and-drop functionality.
+ */
+
+export { Timeline } from "./Timeline";
+export { TimelineControls } from "./TimelineControls";
+export { TimelineRuler } from "./TimelineRuler";
+export { TimelineToolbar } from "./TimelineToolbar";
+export { TimelineWaveform } from "./TimelineWaveform";
+export { Track } from "./Track";
+export { TrackList } from "./TrackList";
+export { Clip } from "./Clip";
+export { ClipDragLayer } from "./ClipDragLayer";
+export { ClipFilmstrip } from "./ClipFilmstrip";
+export { Playhead } from "./Playhead";
+export { GhostTrack } from "./GhostTrack";
+export { EmptyTimelineDropZone } from "./EmptyTimelineDropZone";
