+++ src/components/editor/transform/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Transform Components
+ *
+ * Components for handling clip transformations and overlays.
+ */
+
+export { TransformOverlay } from "./TransformOverlay";
