+++ src/components/editor/viewport/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Viewport Components
+ *
+ * Components for viewport controls, overlays, and transformations.
+ */
+
+export * from "./ViewportControls";
+export { SafeOverlay } from "./SafeOverlay";
